moved _partition function to the stand alone helper function (it is not relevant to the analysis). Modified VCF to Structure method to allow non labelled str file at the end.
