+++ mycorrhiza/analysis/analysis.py
@@ -1,10 +1,8 @@
 import numpy as np
 from typing import List
 from ..dataset.dataset import Dataset
-from os import path
 from ..dataset import nexus
 from ..helper import random_string
-import os
 from tqdm import tqdm
 import random
 from pathos.multiprocessing import Pool
@@ -48,6 +46,32 @@ class Result:
 	def identifiers(self):
 		return self._dataset.identifiers
 
+	def set_q_pred_pops_params(self, predicted_origin, probabilities, classes, test_indices=[], train_indices=[]):
+		'''
+		@TBD
+		'''
+		if not test_indices and not train_indices:
+			raise ValueError("Either test or train indices must be provided.")
+		elif not test_indices:
+			test_indices = [ind for ind in range(len(self.real_populations)) if ind not in train_indices]
+		elif not train_indices:
+			train_indices = [ind for ind in range(len(self.real_populations)) if ind not in test_indices]
+		
+		out_q = []
+		out_origin = []
+		for i in np.sort(test_indices + train_indices):
+			if i in test:
+				out_q.append(probabilities[test.index(i)])
+				out_origin.append(predicted_origin[test.index(i)])
+			else:
+				out_q.append([1 if x == list(classes).index(self.real_populations[i]) else 0 for x in range(len(classes))])
+				out_origin.append(self.real_populations[i])
+
+		out_origin, out_q = np.array(out_origin), np.array(out_q)
+
+		self.set_pred_pops(pred_pops)
+		self.set_q(out_q, classes)
+
 	def set_pred_pops(self, pops: List):
 		self._pred_populations = pops
 
@@ -68,7 +92,7 @@ class Result:
 		self._q_pops = col_pops
 
 	def output_q(self, file_name: str='Q'):
-		out_file = path.join(self._out_path, '{0}.tsv'.format(file_name))
+		out_file = os.path.join(self._out_path, '{0}.tsv'.format(file_name))
 
 		if os.path.isfile(out_file):
 			os.remove(out_file)
@@ -79,7 +103,7 @@ class Result:
 				fout.write('{0}\t{1}\n'.format(iden, '\t'.join(line.astype(str))))
 
 	def output_accuracy(self, file_name: str='acc'):
-		out_file = path.join(self._out_path, '{0}.tsv'.format(file_name))
+		out_file = os.path.join(self._out_path, '{0}.tsv'.format(file_name))
 
 		if os.path.isfile(out_file):
 			os.remove(out_file)
@@ -93,56 +117,6 @@ class Result:
 			fout.write('samples\ttrue\tpred\n')
 			for iden, re, pr in zip(self._dataset.identifiers, self.real_populations, self._pred_populations):
 				fout.write('{0}\t{1}\t{2}\n'.format(iden, re, pr))
-
-
-def _as_nexus_file(data: Dataset, out_path, loci: list):
-
-	temp = random_string(32) + '.nex'
-	temp_file = os.path.join(out_path, temp)
-
-	with open(temp_file, 'w+') as fp:
-		nexus.writeHeader(fp, 'MYCORRHIZA', data.num_samples)
-		if data.is_str:
-			nexus.writeDistancesBlock(fp, data._microsatellite_distances(), data.identifiers)
-			nexus.writeFooter(fp, ['SplitsPostProcess filter=weight value=1E-6'])
-		else:
-			nexus.start_characters_block(fp, len(loci))
-			count = 0
-			with tqdm(total=data.num_samples, desc='Outputting data in Nexus format.') as progress:
-				for sample, genotype in data.iterator():
-					progress.update(1)
-
-					nexus.write_characters_line(fp, count, sample.identifier, np.array(genotype)[loci])
-					count += 1
-			nexus.end_block(fp)
-			nexus.writeFooter(fp, ['chartransform=JukesCantor', 'SplitsPostProcess filter=weight value=1E-6'])
-
-	return temp_file
-
-
-def _partition(data: Dataset, out_path, num_partitions: int, num_loci: int, num_cores: int):
-	if num_loci == 0:
-		loci = range(data.num_loci)
-	else:
-		loci = random.sample(range(data.num_loci), num_loci)
-
-	loci = list(loci)
-
-	if data.diploid:
-		loci + [x + data.num_loci for x in loci]
-
-	parts = np.array_split(loci, num_partitions)
-
-	part_files = [_as_nexus_file(data, out_path, x) for x in parts]
-
-	sn = SplitNetwork()
-
-	def func(file):
-		return sn.execute_nexus_file(file)
-
-	p = Pool(num_cores)
-
-	return p.map(func, part_files)
 
 
 def _r_forests(partitions: list, populations: list, num_splits: int, num_estimators: int, num_cores: int):
