+++ mycorrhiza/dataset/vcf.py
@@ -61,7 +61,7 @@ def _write_line(source_file, out_pointer, identifier, population):
 		out_pointer.write('\n')
 
 
-def vcf_to_structure(path_in: str, path_out: str, populations: list, diploid: bool=True):
+def vcf_to_structure(path_in: str, path_out: str, populations: list=None, diploid: bool=True):
 	"""Converts a VCF file into the STRUCTURE format.
 
 	Args:
@@ -78,8 +78,12 @@ def vcf_to_structure(path_in: str, path_out: str, populations: list, diploid: bo
 
 	if len(identifiers) != len(np.unique(identifiers)):
 		raise LoadingError('Sample identifiers must be unique.')
-	if len(populations) != len(identifiers):
+	if populations == None:
+		populations = ['nan'] * len(identifiers) # in case of unlabeled vcf file
+	elif len(populations) != len(identifiers):
 		raise ValueError('Mismatch between the number of samples and populations labels.')
+	else:
+		pass
 
 	print('Opened VCF file with {0} samples and {1} loci.'.format(len(identifiers), num_loci))
